TFP-4709 sanere lesing fra utgåtte topics
