--- integrasjontjenester/ytelsesvedtak-kafka-poller/src/main/java/no/nav/foreldrepenger/tilbakekreving/hendelser/felles/HendelseReader.java
@@ -1,8 +1,0 @@
-package no.nav.foreldrepenger.tilbakekreving.hendelser.felles;
-
-import no.nav.foreldrepenger.tilbakekreving.kafka.poller.PostTransactionHandler;
-
-public interface HendelseReader {
-
-    PostTransactionHandler hentOgBehandleMeldinger();
-}
