+++ web/src/test/java/no/nav/foreldrepenger/tilbakekreving/web/server/jetty/AvhengigheterTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import no.nav.foreldrepenger.tilbakekreving.avstemming.AvstemmingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.datavarehus.saksstatistikk.SakshendelserEventObserver;
 import no.nav.foreldrepenger.tilbakekreving.fplos.klient.observer.FplosEventObserver;
-import no.nav.foreldrepenger.tilbakekreving.hendelser.felles.YtelsesvedtakHendelsePoller;
+import no.nav.foreldrepenger.tilbakekreving.hendelser.VedtakFattetHendelsePoller;
 import no.nav.foreldrepenger.tilbakekreving.kravgrunnlag.queue.consumer.KravgrunnlagAsyncJmsConsumer;
 import no.nav.vedtak.felles.testutilities.cdi.WeldContext;
 
@@ -25,7 +25,7 @@ public class AvhengigheterTest {
         Assertions.assertThat(CDI.current().select(SakshendelserEventObserver.class).isResolvable()).isTrue();
         Assertions.assertThat(CDI.current().select(FplosEventObserver.class).isResolvable()).isTrue();
         Assertions.assertThat(CDI.current().select(KravgrunnlagAsyncJmsConsumer.class).isResolvable()).isTrue();
-        Assertions.assertThat(CDI.current().select(YtelsesvedtakHendelsePoller.class).isResolvable()).isTrue();
+        Assertions.assertThat(CDI.current().select(VedtakFattetHendelsePoller.class).isResolvable()).isTrue();
         Assertions.assertThat(CDI.current().select(AvstemmingTjeneste.class).isResolvable()).isTrue();
     }
 }
